Do not delete manually image if in production

+++ core/models.py
@@ -2,6 +2,7 @@ import io
 import os
 import uuid
 
+from django.conf import settings
 from django.core.files.images import ImageFile
 from django.db import models
 from django.dispatch import receiver
@@ -45,6 +46,9 @@ class Picture(models.Model):
 
 @receiver(models.signals.post_delete, sender=Picture)
 def auto_delete_picture_file_on_delete(sender, instance, **kwargs):
+    if not settings.DEBUG:
+        return
+
     if instance.picture:
         if os.path.isfile(instance.picture.path):
             os.remove(instance.picture.path)
